# port masked_select from TH to ATen and optimize perf on CPU