@mingfeima 2020-02-13
In order to parallel `masked_select` with TensorIterator, extra meta data which records the `offset` of output tensor is needed, `masked_prefix_sum`. It is calculated with **std::partial_sum** at the moment since PyTorch supports C++14. Once PyTorch upgrades to C++17, we can use **std::exclusive_scan** which is from parallel STL, thus have better performance on CPU. I checked CUDA implementation has used **thrust::exclusive_scan** but CPU counterpart has dependency on C++17.

Below shows the performance compare between **original** and **this pr**, benchmark code available at this [op_bench-py](https://github.com/mingfeima/op_bench-py), to reproduce, `./run.sh masked_select.py`. Test machine is Xeon Skylake 6148 with 2*20 cores @2.40GHz.

The unit is `ms per iteration`, the lower the better.
The benchmark takes `mask = input.ge(0.0)`, so roughly **50%** of the input elements are selected.

#### single socket run

input size | original | this pr | speedup
-- | -- | -- | --
[128 1000] | 0.817 | 0.295 | 2.77
[256 1000] | 1.899 | 0.413 | 4.60
[512 1000] | 3.747 | 0.700 | 5.35
[1024 1000] | 6.696 | 1.238 | 5.41

#### single core run

input size | original | this pr | speedup
-- | -- | -- | --
[128 1000] | 0.812 | 0.703 | 1.16
[256 1000] | 1.620 | 1.421 | 1.14
[512 1000] | 3.246 | 2.823 | 1.15
[1024 1000] | 6.488 | 5.675 | 1.14