@VitalyFedyunin 2020-02-24
```
Feb 24 08:29:38 ======================================================================
Feb 24 08:29:38 ERROR [0.008s]: test_fmod_cpu_float32 (__main__.TestTorchDeviceTypeCPU)
Feb 24 08:29:38 ----------------------------------------------------------------------
Feb 24 08:29:38 Traceback (most recent call last):
Feb 24 08:29:38   File "/opt/conda/lib/python3.6/site-packages/torch/testing/_internal/common_device_type.py", line 197, in instantiated_test
Feb 24 08:29:38     result = test(self, device_arg, dtype)
Feb 24 08:29:38   File "/opt/conda/lib/python3.6/site-packages/torch/testing/_internal/common_device_type.py", line 398, in only_fn
Feb 24 08:29:38     return fn(slf, device, *args, **kwargs)
Feb 24 08:29:38   File "test_torch.py", line 13577, in test_fmod
Feb 24 08:29:38     _test_fmod_with_size_tensor(size, device, dtype)
Feb 24 08:29:38   File "test_torch.py", line 13556, in _test_fmod_with_size_tensor
Feb 24 08:29:38     a = torch.rand(size=size, device=device, dtype=dtype)
Feb 24 08:29:38 RuntimeError: _th_uniform_ not supported on CPUType for Byte
```