@VitalyFedyunin 2020-02-25
> > test_fmod_cpu_float32
> 
> Sorry, it is this related to this one?

Oops, wrong browser tab, sorry