@kurtamohler 2020-02-25
Hey @mingfeima , are you sure your performance comparison between your implementation and the old TH implementation is correct? I was running some measurements on my cuda implementation, and I decided to collect the cpu performance data as well to make sure my changes won't negatively affect yours. Here's the performance I see with a single CPU core:

```
dtype | tensor_size | mask_true_ratio | TH CPU time | ATen cpu (mingfeima) time | speedup mingfeima ATen over TH
-- | -- | -- | -- | -- | --
torch.float64 | 1000 | 0 | 0.0000146 | 0.0001218 | 0.1198686371
torch.float64 | 1000 | 0.25 | 0.0000182 | 0.0001524 | 0.1194225722
torch.float64 | 1000 | 0.5 | 0.0000176 | 0.0001544 | 0.1139896373
torch.float64 | 1000 | 0.75 | 0.0000174 | 0.0001538 | 0.1131339402
torch.float64 | 1000 | 1 | 0.0000172 | 0.0001504 | 0.1143617021
torch.float64 | 10000 | 0 | 0.0000524 | 0.0006164 | 0.08500973394
torch.float64 | 10000 | 0.25 | 0.0000722 | 0.0007122 | 0.101376018
torch.float64 | 10000 | 0.5 | 0.000076 | 0.0007412 | 0.1025364274
torch.float64 | 10000 | 0.75 | 0.000072 | 0.000738 | 0.09756097561
torch.float64 | 10000 | 1 | 0.0000586 | 0.0007264 | 0.08067180617
torch.float64 | 100000 | 0 | 0.000439 | 0.0055718 | 0.07878961915
torch.float64 | 100000 | 0.25 | 0.0006324 | 0.006367 | 0.09932464269
torch.float64 | 100000 | 0.5 | 0.0007926 | 0.0067212 | 0.1179253705
torch.float64 | 100000 | 0.75 | 0.0006554 | 0.0066458 | 0.09861867646
torch.float64 | 100000 | 1 | 0.0004654 | 0.0065112 | 0.07147683991
torch.float64 | 1000000 | 0 | 0.0042768 | 0.0559348 | 0.07646045038
torch.float64 | 1000000 | 0.25 | 0.0063654 | 0.064207 | 0.09913872319
torch.float64 | 1000000 | 0.5 | 0.0083026 | 0.0674632 | 0.1230685766
torch.float64 | 1000000 | 0.75 | 0.0067032 | 0.0665706 | 0.1006930988
torch.float64 | 1000000 | 1 | 0.0050948 | 0.0653446 | 0.07796818712
torch.float64 | 10000000 | 0 | 0.0424076 | 0.577658 | 0.07341298831
torch.float64 | 10000000 | 0.25 | 0.0647738 | 0.6789826 | 0.09539832096
torch.float64 | 10000000 | 0.5 | 0.0916282 | 0.7228486 | 0.1267598775
torch.float64 | 10000000 | 0.75 | 0.0796012 | 0.7180316 | 0.110860302
torch.float64 | 10000000 | 1 | 0.0646096 | 0.7097978 | 0.09102535962
```

My performance measurement script is here: https://github.com/kurtamohler/pytorch-perf-test-scripts/blob/master/masked_select/masked_select-perf.py