@kurtamohler 2020-02-25
Actually, I think your performance measurement script is sufficient. When I run it, I get something that seems to be similar to my measurements, showing a performance decrease.

TH CPU masked_select:
```
input size: [128 1000]; output size: [63941]: time = 1.027 ms
input size: [256 1000]; output size: [127483]: time = 2.225 ms
input size: [512 1000]; output size: [255929]: time = 4.103 ms
input size: [1024 1000]; output size: [511179]: time = 7.888 ms
```

ATen CPU masked_select:
```
input size: [128 1000]; output size: [63941]: time = 5.091 ms
input size: [256 1000]; output size: [127483]: time = 6.737 ms
input size: [512 1000]; output size: [255929]: time = 10.314 ms
input size: [1024 1000]; output size: [511179]: time = 17.129 ms
```



TH CPU masked_select (modified to use only a single core):
```
input size: [128 1000]; output size: [63941]: time = 1.030 ms
input size: [256 1000]; output size: [127483]: time = 2.098 ms
input size: [512 1000]; output size: [255929]: time = 4.241 ms
input size: [1024 1000]; output size: [511179]: time = 8.834 ms
```

ATen CPU masked_select (modified to use only a single core):
```
input size: [128 1000]; output size: [63941]: time = 8.616 ms
input size: [256 1000]; output size: [127483]: time = 17.228 ms
input size: [512 1000]; output size: [255929]: time = 34.527 ms
input size: [1024 1000]; output size: [511179]: time = 69.079 ms
```