@mingfeima 2020-02-26
@kurtamohler What type of CPU are you using? And how about the environment setting?